building-an-app-3/requirements.txt
Flask==3.0.0
google-cloud-datastore==2.15.1
pytz==2023.3.post1
google-auth==2.17.3
requests==2.28.2